requirements.txt
numpy>=1.23
pandas>=1.5
pyyaml>=6.0
cvxpy>=1.3
matplotlib>=3.5
scipy>=1.9
pyarrow>=11.0
pkg_resources>=0.0
# Optional GPU: install cupy matching your CUDA (see instructions). Do NOT put cupy here unconditionally.
